Add support for Amazon Kinesis Client Library 2.x
Removed the use of a state machine for message handling

Removed the state machine that was used for message handling and went
to a simpler dispatch model.

Classes and interfaces have been moved from a single large file to
individual files.

Added a new interface IShardRecordProcessor that replaces shutdown
with specific methods: leaseLost, and shardEnded.

Added options to the Bootstrap project for configuring logback in the
Java KCL.